Issue #53
Refactor grid search tree to its own source file.
